+++ src/getWhitelistedRootPackagesLicenses.js
@@ -1,6 +1,60 @@
 'use strict';
 
-module.exports = function (bundledRootPackages) {
+const {readFile} = require('fs/promises');
+const {join} = require('path');
+
+const yaml = require('js-yaml');
+
+module.exports = async function (
+  bundledRootPackages, packagePath, production
+) {
+  let packageLock;
+  let pnpm, yarn;
+  try {
+    // eslint-disable-next-line max-len -- Long
+    // eslint-disable-next-line import/no-dynamic-require, node/global-require -- Runtime detect
+    packageLock = require(join(packagePath, 'package-lock.json')).packages;
+  } catch (e) {
+    // istanbul ignore next
+    if (e) {
+      /* eslint-disable no-console -- CLI */
+      // istanbul ignore next
+      console.error(
+        'No package-lock.json file found and pnpm-lock.yaml and yarn.lock ' +
+        'files are not currently supported'
+      );
+      /* eslint-enable no-console -- CLI */
+      // istanbul ignore next
+      return [];
+    }
+    // istanbul ignore next
+    try {
+      // istanbul ignore next
+      packageLock = yaml.load(await readFile(
+        join(packagePath, 'pnpm-lock.yaml')
+      )).packages;
+      // istanbul ignore next
+      pnpm = true;
+      // console.log('packageLock', packageLock);
+    // istanbul ignore next
+    } catch (pnpmErr) {
+      // istanbul ignore next
+      packageLock = yaml.load(await readFile(
+        join(packagePath, 'yarn.lock')
+      )).packages;
+      // istanbul ignore next
+      yarn = true;
+      // istanbul ignore next
+      if (yarn) {
+        // Add this condition below as needed when may be ready
+      }
+    }
+  }
+
+  // Todo: Other lock files
+  // yaml.load(await readFile('yarn.lock'));
+  // yaml.load(await readFile('pnpm-lock.yaml'));
+
   return (unflattenedPackages) => {
     const packages = [];
     const flatten = (pkg) => {
@@ -29,17 +83,48 @@ module.exports = function (bundledRootPackages) {
       //   over
       // const isRootDep = pkg.package._requiredBy.includes('#USER');
       // Wasn't able to replicate by `npm i --no-save`
+
+      // Todo: Per https://github.com/npm/rfcs/blob/latest/implemented/0013-no-package-json-_fields.md ,
+      //    this `_requiredBy` is being removed (and doesn't work with pnpm
+      //    or Yarn)
       /* istanbul ignore if */
+      /*
       if (!pkg.package._requiredBy) {
         // May have been installed in `node_modules` but unused
+        console.log('pkg.package', pkg.package);
         return false;
       }
-      const isRootDevDep = pkg.package._requiredBy.includes('#DEV:/');
-      return isRootDevDep &&
-        (bundledRootPackages === true ||
+      const isRootDep = pkg.package._requiredBy.includes('#DEV:/');
+      */
+      // Todo: was `pkg.package` here, but we are checking `package.json` keys
+      //   instead of what the name inside the (GitHub) package name is, e.g.,
+      //   brettz9-missing points to the package with name `license-missing`.
+      const {name, version = pkg.package.version} = pkg;
+      const isRootDep = Object.entries(packageLock).some((
+        [packageKey, value]
+      ) => {
+        return (
+          // istanbul ignore next
+          (pnpm && packageKey === `/${name}/${version}`) ||
+          (
+            !pnpm && packageKey === `node_modules/${name}` &&
+            value.version === version
+          )
+        ) &&
+          (
+            !production || !value.dev
+          );
+      });
+
+      return isRootDep &&
+        (production || bundledRootPackages === true ||
           (Array.isArray(bundledRootPackages) &&
             bundledRootPackages.includes(pkg.name)));
     });
+
+    // console.log('filteredPackages', filteredPackages.map(({name}) => {
+    //   return name;
+    // }));
 
     /**
      * `package.json` info.